requirements.txt
# AI Video Automation - Dependencies
# Python 3.10+ required

# Gemini API (新しいライブラリ)
google-genai>=0.2.0

# YouTube API
google-api-python-client>=2.150.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# HTTP Client
httpx>=0.27.0

# Video Editing (2.x for new API)
moviepy>=2.0.0

# Image Processing
Pillow>=10.4.0

# Environment
python-dotenv>=1.0.1

# CLI & Progress
rich>=13.9.0
click>=8.1.0

# Utilities
requests>=2.32.0

# News Automation
elevenlabs>=1.0.0
pytrends>=4.9.2
newsapi-python>=0.2.7
feedparser>=6.0.10

# Web Scraping
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
lxml>=5.0.0
